updated: some error colors; fixed: error 320, 220
