+++ css/80_app.css
@@ -2572,7 +2572,7 @@ input.key-trap {
 .kr_error_type_41,
 .kr_error_type_42,
 .kr_error_type_43 {
-    color: #894668;
+    color: #fd007f;
 }
 
 .kr_error_type_50 {
@@ -2618,7 +2618,7 @@ input.key-trap {
 }
 
 .kr_error_type_180 {
-    color: #09ef12;
+    color: #01ff0a;
 }
 
 .kr_error_type_190,
@@ -2642,7 +2642,7 @@ input.key-trap {
 .kr_error_type_206,
 .kr_error_type_207,
 .kr_error_type_208 {
-    color: #71f264;
+    color: #fdbf6f;
 }
 
 .kr_error_type_210,
@@ -2659,7 +2659,7 @@ input.key-trap {
 .kr_error_type_230,
 .kr_error_type_231,
 .kr_error_type_232 {
-    color: #5f775c;
+    color: #b15928;
 }
 
 .kr_error_type_270 {
@@ -2684,7 +2684,7 @@ input.key-trap {
 .kr_error_type_296,
 .kr_error_type_297,
 .kr_error_type_298 {
-    color: #d1dce7;
+    color: #a6cee3;
 }
 
 .kr_error_type_310,
@@ -2699,7 +2699,7 @@ input.key-trap {
 }
 
 .kr_error_type_350 {
-    color: #4d719c;
+    color: #ffff99;
 }
 
 .kr_error_type_370 {
@@ -2713,7 +2713,7 @@ input.key-trap {
 .kr_error_type_400,
 .kr_error_type_401,
 .kr_error_type_402 {
-    color: #b20e36;
+    color: #b64f69;
 }
 
 .kr_error_type_410,
