+++ modules/util/keepRight/index.js
@@ -1,2 +1,3 @@
-export { parseErrorDescriptions, clickLink } from './keepRight_error';
 export { errorTypes } from './errorSchema.json';
+export { parseError } from './parse_error';
+export { parseErrorDescriptions, clickLink } from './keepRight_error';
